Since you're publishing to npm - please show this is a fork
Hello

I was confused by this on npm. In github, it's totally clear that its a fork of dungeon-generator. But npm presents them as two separate projects. 
Could you include a note for fellow npm users, to make the project's status clear?

This PR also removes the live demo link, since it is for the other project.

Thanks,
 - Daniel